Adds a timeout to the SSH channel.
This resolves an issue in which a channel.recv() operation can hang when
the host's sshd is alive but netconfd has hung or fallen over.  The TCP
socket and Paramiko Transport do not time out because data is still
flowing, but the NETCONF connection hangs.  We have observed a simple
NETCONF HELLO exchange to hang for over 2.5 hours in the wild despite
the socket timeout being set.

Paramiko channels, it turns out, have their own timeout that is
independdent of the socket/transport timeout.