add: specific timeout to Netconf client RPCs (#47)
* add: specific timeout to Netconf client RPCs

All public RPC functions of class Manager now have a new timeout
default parameter. If not given (or set to None), the Manager's
timeout value is used as before.
If given, the specific timeout is used for this one RPC, the
original timeout value stored in the Manager is not changed.

Moreover, the timeout value given once at construction of the
Manager object can now be changed at runtime.

* fix: black formatting issues

* fix: version numbering

---------

Co-authored-by: Wolfram Burkhardt <wolfram.burkhardt@adtran.com>